Rename directory which contains patches only when needed
For example if the directory is called files or has the name of the package
is better to keep it that way to avoid conflicts or overriding of patches
for a different version of the same recipe.

The renaming is done, however, when the package version is in the name of
the directory. Otherwise paches will not be found.

Signed-off-by: Marius Avram <marius.avram@intel.com>