Improvements on compilation failed
- will try to remove only the patch which caused compilation failure.
The Upstream Status of the patch removed will be mentioned in the
commit message. If the compilation still failed after the faulty
patches have been removed it will add the patches back because most
probably they will have to be changed for the structure of the source
code of the package.

- it will also try to automatically add new files which were not
shipped in the final package.

Signed-off-by: Marius Avram <marius.avram@intel.com>