Send emails for a list of recipes
If the --send-emails or -e command line argument is active and a single
or a list of recipes are given to the script, emails with the results
and patches will be sent to their maintainers or to the overriden
addresses from the configuration file.

Also the status email will be sent in the case a selection of
at least two recipes has been given to the script for upgrading.

Signed-off-by: Marius Avram <marius.avram@intel.com>
Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>