testimage.py: remove all iterations (over machines, over failing recipes)
This adds enormously to run time, and is not particularly necessary
when running locally. If something fails, it's best to let the maintainer sort
it out as soon as possible, rather than try to guess what needs to be excluded.

Also, do not build a separate ptest image; ptests will be added
to the standard integration testimage.

Signed-off-by: Alexander Kanavin <alexander.kanavin@linux.intel.com>