Improve README and sample config file.
Specifically, add more information about various config settings,
add a few important points to README, and remove config file specifics
from README, so that there's only one source for what the settings are
(the sample config file).

Also comment out everything from the config file, so that those in a
hurry can take it into use immediately. Most settings are specific to
a maintainer, and so can't be provided in a 'generic' way.

Signed-off-by: Alexander Kanavin <alexander.kanavin@linux.intel.com>