Add a commit_revert_policy configuration setting
This helps in particular when a package upgrade builds fine, but then
causes other packages to fail. Depending on how close to the root
of dependency tree the failing packages are, this can badly subvert
the unattended upgrade process. So for that case, the better approach
is to revert every upgrade commit, so that each upgrade is effectively
built against master, which makes it less likely to fail.

Signed-off-by: Alexander Kanavin <alexander.kanavin@linux.intel.com>