upgradehelper: Replace do_checkpkg usage with oe.recipeutils function call
The code in distrodata.bbclass related to the do_checkpkg task is rather
dated, has holes in it (ignoring some recipes) and has horrible locking
and csv related issues.

We should use modern APIs such as tinfoil to make the calls we need directly
against bitbake, cutting out the middleman and clarifing the code.

This change uses a lib/oe/recipeutils function call to remove the need
to use checkpkg at all.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>